Add missing API root for get token
